requirements.txt
Flask==1.0.2
gunicorn==19.3.0
google.cloud.bigquery
google-cloud-bigquery